# ENH: faster get_log_path()