Related to gh-606.

* this is a precursor to using `get_log_path()` ubiquitously
in the test suite, which itself is a precursor to being
able to run the test suite from any directory (including
a user/developer with a `spack` install of `py-darshan`, which
is our primary planned vehicle for initial delivery of the new
report capability)

* achieve nearly constant-time log filepath retrieval
using a dictionary/hash table

* the speedups on repeated calls, the most important scenario
for running in `pytest`, are several orders of magnitude faster using
the new benchmark in gh-610:

`asv continuous pydarshan-devel treddy_get_log_path_dict -e -b
"time_get_log_path_repeat`

```
       before           after         ratio
     [6d0a48e0]       [47e9a462]
     <pydarshan-devel>       <treddy_get_log_path_dict>
-        97.6±2ms        962±100ns     0.00
         log_utils.GetLogPath.time_get_log_path_repeat(1, 'dxt.darshan')
-         974±7ms       3.14±0.1μs     0.00
          log_utils.GetLogPath.time_get_log_path_repeat(10,
'dxt.darshan')
-         2.43±0s       7.31±0.3μs     0.00
          log_utils.GetLogPath.time_get_log_path_repeat(25,
'dxt.darshan')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```

* the new tests from gh-609 also seem to pass with this branch,
though it would probably be prudent to review/fix/merge that
first..